filter product by category

+++ lib/providers/cart.dart
@@ -4,7 +4,7 @@ class CartItem {
   final String id;
   final String title;
   final int quantity;
-  final double price;
+  final int price;
 
   CartItem({
     @required this.id,
@@ -24,15 +24,15 @@ class Cart with ChangeNotifier {
     return _items.length;
   }
 
-  double get totalAmount {
-    var total = 0.0;
+  int get totalAmount {
+    var total = 0;
     _items.forEach((key, cartItem) {
       total += cartItem.quantity * cartItem.price;
     });
     return total;
   }
 
-  void addItem(String productId, String title, double price) {
+  void addItem(String productId, String title, int price) {
     if (_items.containsKey(productId)) {
       _items.update(
           productId,
